+++ slash/resuming.py
@@ -117,8 +117,10 @@ def get_last_resumeable_session_id():
         return session_id.session_id
 
 
-def get_tests_to_resume(session_id):
+def get_tests_from_previous_session(session_id, get_successful_tests=False):
     returned = []
+    if get_successful_tests:
+        return get_tests_from_remote_session(session_id, get_successful=True)
     with connecting_to_db() as conn:
          # pylint: disable=no-member
         session_metadata = conn.query(SessionMetadata).filter(SessionMetadata.session_id == session_id).first()
@@ -134,11 +136,11 @@ def get_tests_to_resume(session_id):
                 returned.append(new_entry)
     if not returned:
         _logger.debug('No local entry for session {0}, searching remote session'.format(session_id))
-        returned = resume_remote_session(session_id)
+        returned = get_tests_from_remote_session(session_id)
     return returned
 
 
-def resume_remote_session(session_id):
+def get_tests_from_remote_session(session_id, get_successful=False):
     active_plugins = manager.get_active_plugins()
     backslash_plugin = active_plugins.get('backslash', None)
     if not backslash_plugin:
@@ -147,7 +149,7 @@ def resume_remote_session(session_id):
         raise CannotResume("Backslash plugin doesn't support remote resuming")
     if not backslash_plugin.is_session_exist(session_id):
         raise CannotResume("Could not find resume data for session {0}".format(session_id))
-    remote_tests = backslash_plugin.get_tests_to_resume(session_id)
+    remote_tests = backslash_plugin.get_tests_to_resume(session_id, get_successful)
     returned = [ResumedTestData(test.info['file_name'], test.info['name'], test.variation) for test in remote_tests]
     return returned